+++ changelog.md
@@ -1,5 +1,8 @@
 ## Changelog
 
+### Unreleased
+- Command Line
+  - fixed a bug where absolute paths were treated as relative
 ### 0.5.0 - 2021-05-20
 - Command Line
   - external plugins can be specified in the config file via pluginOpts.path